requirements.txt
numpy
pandas
matplotlib
pyyaml
networkx
scipy
pytest
